requirements.txt
streamlit==1.54.0
google-generativeai==0.8.6
pandas==2.3.3
python-dotenv==1.2.2
